For issue #60.
Moves version related code out from settings files and into their own versions files.
