+++ PlayRho/Common/Version.cpp
@@ -19,8 +19,34 @@
  * 3. This notice may not be removed or altered from any source distribution.
  */
 
-#include <PlayRho/Common/Settings.hpp>
+#include <PlayRho/Common/Version.hpp>
+#include <PlayRho/Common/Templates.hpp>
+#include <PlayRho/Common/RealNum.hpp>
+
+#include <cstdio>
+#include <cstdarg>
+#include <cstdlib>
+#include <typeinfo>
+#include <sstream>
 
 namespace playrho {
-
+    
+    Version GetVersion() noexcept
+    {
+        return Version{0, 9, 0};
+    }
+    
+    std::string GetBuildDetails() noexcept
+    {
+        std::stringstream stream;
+        stream << "asserts=";
+#ifdef NDEBUG
+        stream << "off";
+#else
+        stream << "on";
+#endif
+        stream << ", Real='" << GetTypeName<Real>() << "'";
+        return stream.str();
+    }
+        
 } // namespace playrho
